Split db_access up.
